Separated plotting tools
