+++ src/ConstructiveSolidGeometry/plotting/SurfacePrimitives/SurfacePrimitives.jl
@@ -1,0 +1,7 @@
+include("Plane.jl")
+include("ConalPlane.jl")
+include("CylindricalAnnulus.jl")
+include("ConeMantle.jl")
+include("ToroidalAnnulus.jl")
+include("TorusMantle.jl")
+#include("Sphere.jl")
